+++ src/simulation/weights.jl
@@ -221,7 +221,26 @@ function hostWeights!(host_idx::Int64, population::Population, model::Model)
     end
 end
 
-# Intra-Model (Event) level: None needed
+# Intra-Model (Event) level:
+function updatePopulationContactWeightReceiveMatrix!(pop_idx_1::Int64, pop_idx_2::Int64, change::Float64, model::Model)
+    model.population_contact_weights_receive[pop_idx_2, pop_idx_1] += change
+    model.population_contact_weights_receive_sums[pop_idx_1] += change
+    propagateWeightChanges!(
+        model.populations[pop_idx_1].parameters.base_coefficients[CONTACT] *
+        model.populations[pop_idx_1].contact_sum * change,
+        model.populations[pop_idx_1], CONTACT, model
+    )
+end
+
+function updatePopulationTransitionWeightReceiveMatrix!(pop_idx_1::Int64, pop_idx_2::Int64, change::Float64, model::Model)
+    model.population_transition_weights_receive[pop_idx_2, pop_idx_1] += change
+    model.population_transition_weights_receive_sums[pop_idx_1] += change
+    propagateWeightChanges!(
+        model.populations[pop_idx_1].parameters.base_coefficients[TRANSITION] *
+        model.populations[pop_idx_1].transition_sum * change,
+        model.populations[pop_idx_1], TRANSITION, model
+    )
+end
 
 # Rate level: None needed
 
@@ -234,35 +253,67 @@ end
 
 function propagateWeightChanges!(change::Float64, population::Population, evt::Int64, model::Model)
     model.population_weights[evt, model.population_dict[population.id]] += change
-
-    if evt == CONTACT && population.total_hosts > 0
-        population.contact_sum += change
-        change = change * population.receive_contact_sum / population.total_hosts
-        # elseif evt == INTER_POPULATION_CONTACT
-        #     population.inter_population_contact_sum += change
-    end
-
     propagateWeightChanges!(change, evt, model)
 end
 
 function propagateWeightChanges!(change::Float64, host_idx::Int64, population::Population, evt::Int64, model::Model)
     population.host_weights[evt, host_idx] += change
 
+    if population.total_hosts > 0
+        if evt == CONTACT
+            population.contact_sum += change
+            change = change * model.population_contact_weights_receive_sums[model.population_dict[population.id]]
+        elseif evt == TRANSITION
+            population.transition_sum += change
+            change = change * model.population_transition_weights_receive_sums[model.population_dict[population.id]]
+        end
+    else
+        if evt == CONTACT
+            population.contact_sum = 0.0
+        elseif evt == TRANSITION
+            population.transition_sum = 0.0
+        end
+        change = -model.population_weights[evt, model.population_dict[population.id]]
+    end
+
     propagateWeightChanges!(population.parameters.base_coefficients[evt] * change, population, evt, model)
 end
 
 function propagateWeightReceiveChanges!(change::Float64, population::Population, evt::Int64, model::Model)
     model.population_weights_receive[evt-CHOICE_MODIFIERS[1]+1, model.population_dict[population.id]] += change
-
     model.population_weights_receive_sums[evt-CHOICE_MODIFIERS[1]+1] += change
 
-    if evt == RECEIVE_CONTACT && evt < NUM_COEFFICIENTS
-        population.receive_contact_sum += change
-        propagateWeightChanges!(
-            change * population.contact_sum / population.total_hosts,
-            population, CONTACT, model
-        )
-        #TODO: update inter-pop contacts
+    if evt == RECEIVE_CONTACT
+        for p in 1:length(model.populations)
+            change_p = (
+                change * model.populations[p].population_contact_coefficients[model.population_dict[population.id]] /
+                max(population.total_hosts, 1)
+                )
+                # Contact rates assume scaling area: large populations are equally
+                # dense as small ones, so contact is constant (divide by total hosts).
+                # If you don't want this to happen, modify each population's
+                # receive contact coefficient accordingly.
+            model.population_contact_weights_receive[model.population_dict[population.id], p] += change_p
+            model.population_contact_weights_receive_sums[p] += change_p
+            propagateWeightChanges!(
+                change_p * model.populations[p].contact_sum,
+                model.populations[p], CONTACT, model
+            )
+        end
+    elseif evt == RECEIVE_TRANSITION
+        for p in 1:length(model.populations)
+            change_p = (
+                change * model.populations[p].population_transition_coefficients[model.population_dict[population.id]] *
+                population.population_transition_coefficients_receive[p] / max(population.total_hosts, 1)
+                )
+                # Transition receive weights are assumed to be independent of population.
+            model.population_transition_weights_receive[model.population_dict[population.id], p] += change_p
+            model.population_transition_weights_receive_sums[p] += change_p
+            propagateWeightChanges!(
+                change_p * model.populations[p].transition_sum,
+                model.populations[p], TRANSITION, model
+            )
+        end
     end
 end